fix(practices): Stop double-counting collected revenue and unblock client bundle
Two issues found by CI + Codex review:

1. Collected revenue was double-counted: the loader summed
   Claim.paidAmountCents aggregates AND iterated Payment.amountCents on
   top, so any org with posted payments had its "Collected" KPI inflated
   roughly 2x on both the card and the fleet summary. Claim.paidAmountCents
   already mirrors the sum of payment rows, so the payments query is
   redundant. Dropped the Payment.findMany batch entirely — one fewer
   round-trip, no double-count. Subtitle reworded "Sum of payments" →
   "Posted to claims" to reflect the source.

2. next build failed because practice-card.tsx ("use client") imported
   types and humanize helpers from loaders.ts, which transitively pulls
   "server-only" via super-admin-bootstrap. Split the pure types +
   presentation helpers into a new types.ts so the client component
   doesn't drag the server-only chain into the client bundle. loaders.ts
   re-exports the same names so external callers don't need to update.
